+++ database/database_test.go
@@ -31,7 +31,7 @@ func (suite *DatabaseTestSuite) SetupSuite() {
 
 func (suite *DatabaseTestSuite) TestBuildConnectionOptions() {
 	suite.Equal("goyave:secret@(127.0.0.1:3306)/goyave?charset=utf8&parseTime=true&loc=Local", buildConnectionOptions("mysql"))
-	suite.Equal("host=127.0.0.1 port=3306 user=goyave dbname=goyave password=secret options='charset=utf8&parseTime=true&loc=Local'", buildConnectionOptions("postgres"))
+	suite.Equal("host=127.0.0.1 port=3306 user=goyave dbname=goyave password=secret charset=utf8&parseTime=true&loc=Local", buildConnectionOptions("postgres"))
 	suite.Equal("goyave", buildConnectionOptions("sqlite3"))
 	suite.Equal("sqlserver://goyave:secret@127.0.0.1:3306?database=goyave&charset=utf8&parseTime=true&loc=Local", buildConnectionOptions("mssql"))