+++ router_test.go
@@ -2,6 +2,7 @@ package goyave
 
 import (
 	"io/ioutil"
+	"net/http"
 	"net/http/httptest"
 	"testing"
 
@@ -94,7 +95,8 @@ func (suite *RouterTestSuite) TestStaticHandler() {
 	handler = staticHandler("config", false)
 	handler(response, request)
 	result = response.ResponseWriter.(*httptest.ResponseRecorder).Result()
-	suite.Equal(404, result.StatusCode)
+	suite.Equal(200, result.StatusCode) // Not written yet
+	suite.Equal(404, response.GetStatus())
 
 	body, err = ioutil.ReadAll(result.Body)
 	if err != nil {
@@ -149,6 +151,20 @@ func (suite *RouterTestSuite) TestRequestHandler() {
 	suite.Equal(0, len(body))
 	suite.True(suite.middlewareExecuted)
 	suite.middlewareExecuted = false
+
+	writer = httptest.NewRecorder()
+	router = newRouter()
+	router.requestHandler(writer, rawRequest, func(response *Response, request *Request) {
+		response.Status(http.StatusNotFound)
+	}, validation.RuleSet{})
+
+	result = writer.Result()
+	body, err = ioutil.ReadAll(result.Body)
+	if err != nil {
+		panic(err)
+	}
+	suite.Equal(404, result.StatusCode)
+	suite.Equal("{\"error\":\""+http.StatusText(404)+"\"}\n", string(body))
 }
 
 func (suite *RouterTestSuite) TestCORS() {
@@ -176,6 +192,125 @@ func (suite *RouterTestSuite) TestCORS() {
 	})
 	rawRequest := httptest.NewRequest("GET", "/cors", nil)
 	router.requestHandler(writer, rawRequest, func(response *Response, request *Request) {}, validation.RuleSet{})
+}
+
+func (suite *RouterTestSuite) TestPanicStatusHandler() {
+	request, response := createRouterTestRequest("/uri")
+	response.err = "random error"
+	panicStatusHandler(response, request)
+	result := response.ResponseWriter.(*httptest.ResponseRecorder).Result()
+	suite.Equal(500, result.StatusCode)
+}
+
+func (suite *RouterTestSuite) TestErrorStatusHandler() {
+	request, response := createRouterTestRequest("/uri")
+	response.Status(404)
+	errorStatusHandler(response, request)
+	result := response.ResponseWriter.(*httptest.ResponseRecorder).Result()
+	suite.Equal(404, result.StatusCode)
+	suite.Equal("application/json", result.Header.Get("Content-Type"))
+
+	body, err := ioutil.ReadAll(result.Body)
+	if err != nil {
+		panic(err)
+	}
+	suite.Equal("{\"error\":\""+http.StatusText(404)+"\"}\n", string(body))
+}
+
+func (suite *RouterTestSuite) TestStatusHandlers() {
+	rawRequest := httptest.NewRequest("GET", "/uri", nil)
+	writer := httptest.NewRecorder()
+	router := newRouter()
+	router.StatusHandler(func(response *Response, request *Request) {
+		response.String(http.StatusInternalServerError, "An unexpected panic occurred.")
+	}, http.StatusInternalServerError)
+	router.requestHandler(writer, rawRequest, func(response *Response, request *Request) {
+		panic("Panic")
+	}, validation.RuleSet{})
+
+	result := writer.Result()
+	body, err := ioutil.ReadAll(result.Body)
+	if err != nil {
+		panic(err)
+	}
+	suite.Equal(500, result.StatusCode)
+	suite.Equal("An unexpected panic occurred.", string(body))
+
+	// On subrouters
+	subrouter := router.Subrouter("/sub")
+	writer = httptest.NewRecorder()
+	router = newRouter()
+	subrouter.requestHandler(writer, rawRequest, func(response *Response, request *Request) {
+		panic("Panic")
+	}, validation.RuleSet{})
+
+	result = writer.Result()
+	body, err = ioutil.ReadAll(result.Body)
+	if err != nil {
+		panic(err)
+	}
+	suite.Equal(500, result.StatusCode)
+	suite.Equal("An unexpected panic occurred.", string(body))
+
+	// Multiple statuses
+	writer = httptest.NewRecorder()
+	subrouter.StatusHandler(func(response *Response, request *Request) {
+		response.String(response.GetStatus(), http.StatusText(response.GetStatus()))
+	}, 400, 404)
+	subrouter.requestHandler(writer, rawRequest, func(response *Response, request *Request) {
+		response.Status(400)
+	}, validation.RuleSet{})
+
+	result = writer.Result()
+	body, err = ioutil.ReadAll(result.Body)
+	if err != nil {
+		panic(err)
+	}
+	suite.Equal(400, result.StatusCode)
+	suite.Equal(http.StatusText(400), string(body))
+
+	writer = httptest.NewRecorder()
+	subrouter.requestHandler(writer, rawRequest, func(response *Response, request *Request) {
+		response.Status(404)
+	}, validation.RuleSet{})
+
+	result = writer.Result()
+	body, err = ioutil.ReadAll(result.Body)
+	if err != nil {
+		panic(err)
+	}
+	suite.Equal(404, result.StatusCode)
+	suite.Equal(http.StatusText(404), string(body))
+}
+
+func (suite *RouterTestSuite) TestMuxStatusHandler() {
+	router := newRouter()
+
+	recorder := httptest.NewRecorder()
+	router.muxRouter.NotFoundHandler.ServeHTTP(recorder, httptest.NewRequest("GET", "/", nil))
+	result := recorder.Result()
+	suite.Equal(404, result.StatusCode)
+
+	recorder = httptest.NewRecorder()
+	router.muxRouter.MethodNotAllowedHandler.ServeHTTP(recorder, httptest.NewRequest("GET", "/", nil))
+	result = recorder.Result()
+	suite.Equal(405, result.StatusCode)
+
+	config.Set("debug", false)
+	writer := httptest.NewRecorder()
+	router = newRouter()
+	router.requestHandler(writer, httptest.NewRequest("GET", "/uri", nil), func(response *Response, request *Request) {
+		panic("Panic")
+	}, validation.RuleSet{})
+
+	result = writer.Result()
+	body, err := ioutil.ReadAll(result.Body)
+	if err != nil {
+		panic(err)
+	}
+	suite.Equal(500, result.StatusCode)
+	suite.Equal("{\"error\":\""+http.StatusText(http.StatusInternalServerError)+"\"}\n", string(body))
+	config.Set("debug", true)
 }
 
 func TestRouterTestSuite(t *testing.T) {